% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data}
\alias{data}
\title{R6 Data Class (base)}
\description{
R6 Data Class (base)

R6 Data Class (base)
}
\examples{
loDataObj = data$new(arFormula=formula(mpg~.), arData=mtcars)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{data$new()}}
\item \href{#method-get_data}{\code{data$get_data()}}
\item \href{#method-get_param_names}{\code{data$get_param_names()}}
\item \href{#method-get_param_data}{\code{data$get_param_data()}}
\item \href{#method-get_data_vector}{\code{data$get_data_vector()}}
\item \href{#method-is_result_name}{\code{data$is_result_name()}}
\item \href{#method-clone}{\code{data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new `data` object.
This class is intended to be used by developers of this library only.
Used for managing the internal model development process only.
There should be no need for the users of this library to use this class directly.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data$new(arFormula = NA, arData = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{arFormula}}{Formula to use. Currently only supports
(1) something like DV ~ .
(2) one DV only}

\item{\code{arData}}{Dataframe containing the Independent Variables (IV) the Dependent Variable (DV)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `data` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\subsection{Method \code{get_data()}}{
Gets the underlying data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the underlying dataframe (in expanded model format for categorical variables)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_param_names"></a>}}
\subsection{Method \code{get_param_names()}}{
Gets the independent variable names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data$get_param_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the column names of all the parameters (independent variables) in the model
(in expanded model format for categorical variables)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_param_data"></a>}}
\subsection{Method \code{get_param_data()}}{
Gets the independent variable values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data$get_param_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the parameter values (independent variables) in the model
(in expanded model format for categorical variables)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_vector"></a>}}
\subsection{Method \code{get_data_vector()}}{
Gets a particular column from the underlying data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data$get_data_vector(ar)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ar}}{Column name to fetch}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a certain column from the underlying data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_result_name"></a>}}
\subsection{Method \code{is_result_name()}}{
Checks if a string matches the independent variable name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data$is_result_name(ar)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ar}}{Name to check against the underlying target variable name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns true if the passed argument is the name of the target variable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
