\name{get_peakavg}
\alias{get_peakavg}
\title{Calculate peak flow durations}
\description{get the peak average flows for many flow records.}
\usage{get_peakavg(d, flowduration = c(1, 3, 7, 15, 30, 60, 90, 120, 
    183))}
\arguments{
  \item{d}{a dataframe generated by get_forecast() or get_waterdata().}
  \item{flowduration}{a vector of window sizes for calculating moving averages.}
}
\details{the function uses the get_flowavg() function in package 'zoo' to
compute flow duration timeseries. The maximum value of each averaged
timeseries is the peak flow for that flow duration.}
\value{a dataframe.}

\author{Michael Koohafkan}




\examples{
data(esptraces)
esp.ahoc = subset_forecast(esptraces, location='AHOC1')
get_peakavg(esp.ahoc, c(1,7,15,30,60))
}
