% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqMaxModGP.R
\name{MaxMod}
\alias{MaxMod}
\title{MaxMod algorithm for \code{"lineqGP"} and \code{"lineqAGP"} Models}
\usage{
MaxMod(
  model,
  xtest,
  tol = 1e-04,
  max_iter = 10 * ncol(model$x),
  reward_new_knot = tol,
  reward_new_dim = 1e-09,
  print_iter = FALSE,
  nClusters = 1,
  save_history = FALSE
)
}
\arguments{
\item{model}{an object with class \code{lineqGP} or \code{lineqAGP}}

\item{xtest}{test data for assessing the modification of the MAP estimate}

\item{tol}{a number corresponding to the tolerance of algorithm.
The algorithm stops if the MaxMod criterion < tol}

\item{max_iter}{an integer corresponding to number of iterations}

\item{reward_new_knot}{a number corresponding to the reward of adding a new knot in an existing dimension}

\item{reward_new_dim}{a number corresponding to the reward of adding a new dimension}

\item{print_iter}{a logical variable to print results at each iteration}

\item{nClusters}{an integer corresponding to the number of clusters}

\item{save_history}{a logical variable to save the model at each iteration}
}
\value{
an object with class \code{lineqGP} or \code{lineqAGP} containing the resulting model
}
\description{
A wrapper function for the maximum Modification algorithm.
}
\author{
A. F. Lopez-Lopera
}
