% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqMaxModGP.R
\name{MaxModCriterion}
\alias{MaxModCriterion}
\title{MaxMod criterion for \code{"lineqGP"} Models}
\usage{
MaxModCriterion(
  model,
  ulist,
  xtest,
  activeDim,
  activeDim_IdxSeq,
  idx_add,
  reward_new_dim = 1e-06,
  reward_new_knot = 1e-06,
  iter = 1,
  pred
)
}
\arguments{
\item{model}{an object with class \code{lineqGP}}

\item{ulist}{a list with the location of the knots}

\item{xtest}{test data for assessing the modification of the MAP estimate}

\item{activeDim}{a sequence containing the active dimensions}

\item{activeDim_IdxSeq}{a sequence containing the indices of the actived input dimensions}

\item{idx_add}{index of the dimension to be activated}

\item{reward_new_dim}{a number corresponding to the reward of adding a new dimension}

\item{reward_new_knot}{a number corresponding to the reward of adding a new knot in an existing dimension}

\item{iter}{an integer corresponding to the iteration of the MaxMod algorithm}

\item{pred}{an object with class \code{lineqGP} containing the predictive model}
}
\value{
the value of the knot that minimize the MaxMod criterion and the 
value of the objective function
}
\description{
MaxMod criterion used to add a new knot or a new active dimension (see Bachoc et al., 2020)
}
\references{
F. Bachoc, A. F. Lopez-Lopera, and O. Roustant (2020),
"Sequential construction and dimension reduction of Gaussian processes under inequality constraints".
\emph{ArXiv e-prints}
<arXiv:2009.04188>
}
\author{
A. F. Lopez-Lopera
}
