% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{pivot}
\alias{pivot}
\title{Launch PIVOT}
\usage{
pivot()
}
\description{
Launch PIVOT in the default browser.
}
\details{
This program is based on a collection of R scripts written by members of Junhyong Kim Lab at University of Pennsylvania.
 
Its goal is to facilitate fast and interactive RNA Sequencing data analysis and visualization.

You may not use this file except in compliance with the Kim Lab License located at
 
\url{http://kim.bio.upenn.edu/software/LICENSE}
}
\author{
Qin Zhu \email{qinzhu@outlook.com},
Stephen Fisher \email{safisher@sas.upenn.edu},
Hannah Dueck \email{hdueck@mail.med.upenn.edu},
Sarah Middleton \email{s.a.middlet@gmail.com},
Mugdha Khaladkar \email{mugdhak@mail.med.upenn.edu},
Young-Ji Na \email{youngji@sas.upenn.edu},
Junhyong Kim \email{junhyong@sas.upenn.edu}

Maintainer: Junhyong Kim \email{junhyong@sas.upenn.edu}
}

