% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineqGPutils.R
\name{plot.lineqAGP}
\alias{plot.lineqAGP}
\title{Plot for the \code{"lineqAGP"} S3 Class}
\usage{
\method{plot}{lineqAGP}(x, y, ...)
}
\arguments{
\item{x}{an object with \code{"lineqAGP"} S3 class.}

\item{y}{not used.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Plot with the \code{"lineqAGP"} model.
}
\description{
Plot for the \code{"lineqAGP"} S3 class.
See \code{\link{plot.lineqGP}} for more details.
}
\seealso{
\code{\link{ggplot.lineqGP}}, \code{\link{plot}}
}
\author{
A. F. Lopez-Lopera
}
