\name{plot_all_traces}
\alias{plot_all_traces}
\title{Plot ESP traces.}
\description{plot traces in dataset, as well as mean and median flows.}
\usage{plot_all_traces(d)}
\arguments{
  \item{d}{a dataframe generated by get_forecast() or split_by_*year(get_waterdata() ).}
}

\value{a ggplot plot of all traces, the mean flow, and the median flow.}

\author{Michael Koohafkan}




\examples{
data(esptraces)
plot_all_traces(subset_forecast(esptraces, location='AHOC1'))

data(hspftraces)
plot_all_traces(split_by_wateryear(hspftraces))
}
