\name{plot_countexc}
\alias{plot_countexc}
\title{Box plot and density}
\description{Create box plots for multiple flow durations}
\usage{plot_countexc(d, flowexceeds)}
\arguments{
  \item{d}{flow data generated by get_forecast() or to_*year()}
  \item{flowexceeds}{vector of flow magnitudes for exceedance counts}
}

\value{a ggplot box plot  }

\author{Michael Koohafkan}




\examples{
data(hspftraces)
hspf.wyear = split_by_wateryear(hspftraces)
plot_countexc(hspf.wyear, flowexceeds=c(150, 200, 500))
}
