\name{plot_trace}
\alias{plot_trace}
\title{Plot a single ESP trace.}
\description{plot an individual trace.}
\usage{plot_trace(d, colname)}
\arguments{
  \item{d}{a dataframe generated by get_forecast() or split_by_*year(get_waterdata() ).}
  \item{colname}{the column name of the trace to be plotted.}
}

\value{a ggplot plot.}

\author{Michael Koohafkan}




\examples{
data(esptraces)
plot_trace(esptraces, "AHOC1.2000")

# can also use this function to plot raw timeseries
data(usgstraces)
require(ggplot2)
plot_trace(usgstraces, na.omit('flow')) + scale_x_date('', breaks=scales::date_breaks("years"), labels=scales::date_format("\%Y")) +theme (axis.text.x = element_text(angle=45, hjust=1))
}
