\name{plp3}
\alias{plp3}
\title{Distribution function of the (log) Pearson Type III distribution}
\description{wrapper function for the (log) Pearson type III distribution defined in
package lmomco. Returns NaN vector instead of NULL if parameters are invalid.}
\usage{plp3(q, mu, sigma, gamma)}
\arguments{
  \item{q}{
}
  \item{mu}{see lmomco::cdfpe3}
  \item{sigma}{see lmomco::cdfpe3}
  \item{gamma}{see lmomco::cdfpe3}
}



\author{Michael Koohafkan}





