\name{predict_exceedance}
\alias{predict_exceedance}
\title{Calculate exceedance curves.}
\description{generate fitted exceedance curves for plotting.}
\usage{predict_exceedance(flowdata, parameterdata, distr)}
\arguments{
  \item{flowdata}{flows for which exceedance probabilities. Column names must
match columns in argument parameterdata.}
  \item{parameterdata}{dataframe of parameters as generated by get_parameters().}
  \item{distr}{The distribution corresponding to the parameter estimates. Can be either
log Pearson III ('lp3') or 3-parameter lognormal ('ln3').}
}

\value{a dataframe for plotting exceedance curves for each flow duration. Rows
correspond to probability values contained in column 'p.exceedance'. Column 
names are the flow durations.}

\author{Michael Koohafkan}




\examples{
data(workedexample)
# compare predictions using L-moment and refined estimates
predict_exceedance(usgs.fd, usgs.lp3, distr='lp3')
predict_exceedance(usgs.fd, usgs.lp3.ref, distr='lp3')

# predict new flow values
newflows = data.frame(c(1000, 2000), c(1000, 2000))
names(newflows) = c('1.day.flow', '3.day.flow')
predict_exceedance(newflows, usgs.lp3.ref, distr='lp3')
}
