\name{split_by_calendaryear}
\alias{split_by_calendaryear}
\title{Split a timeseries by calendar year.}
\description{splits a timeseries downloaded using the waterData package by calendar year.}
\usage{split_by_calendaryear(d)}
\arguments{
  \item{d}{a dataframe generated by get_waterdata().}
}

\value{a dataframe with each column containing data from each 
calendar year in the record. Missing values are NA.}

\author{Michael Koohafkan}




\examples{
data(usgstraces)
split_by_calendaryear(usgstraces)

data(hspftraces)
split_by_calendaryear(hspftraces)
}
