\name{subset_waterdata}
\alias{subset_waterdata}
\title{Subset downloaded USGS flow data.}
\description{subset waterdata by interval.}
\usage{subset_waterdata(d, sc = c("WYD", "GMT"), startdate = NULL, enddate = NULL)}
\arguments{
  \item{d}{a dataset generated by split_by_wateryear() or split_by_calendaryear().}
  \item{sc}{search column to use for subsetting by date.}
  \item{startdate}{the first day of the interval, either 'MM-DD' string or 3-digit WYD.}
  \item{enddate}{the last day of the interval, same format as startdate.}
}

\value{a subset of the original dataframe.}

\author{Michael Koohafkan}




\examples{
data(usgstraces)
usgs.wyear = split_by_wateryear(usgstraces)
subset_waterdata(usgs.wyear, sc='GMT', startdate='11-01', enddate='03-31')
subset_waterdata(usgs.wyear, sc='WYD', startdate=32, enddate=182)

data(hspftraces)
hspf.cyear = split_by_calendaryear(hspftraces)
subset_waterdata(hspf.cyear, sc='GMT', startdate='01-01', enddate='06-30')

## Not run:
## # will produce errors
## subset_waterdata(hspf.cyear, sc='WYD', startdate=32, enddate=182)
## subset_waterdata(arroyohondo.wyear, sc='GMT', startdate='09-25', enddate='10-05')
## subset_waterdata(hspf.cyear, sc='GMT', startdate='12-25', enddate='01-05')
## End(Not run)
}
