\name{usgstraces}
\docType{data}
\alias{usgstraces}

\title{Sample ESP trace data}
\description{Sample USGS trace data. Generated by calling get_waterdata() on 5/25/2014.}
\usage{
data(usgstraces)
}
\format{
\enumerate{
  \item GMT. Date in YYYY-MM-DD format.
  \item WYD. Date in water-year day format, i.e. 2014-10-01 is 2014001.
  \item flow. Daily average flow in cubic feet per second. 
}
}
\examples{
data(usgstraces)
summary(usgstraces)
}
\keyword{datasets}
