\name{pcalg-internal}
%
\alias{adj.check}
\alias{allDags}
\alias{amat2dag}
\alias{ci.test}
\alias{find.sink}
\alias{flipEdges}
\alias{udag2pdagRelaxed}
\alias{udag2pdagSpecial}
\alias{legal.psep}
\alias{lm.cov}
\alias{pcAlgo.Perfect}
\alias{reach}
\alias{trueCov}
\alias{wgtMatrix}
\alias{check.new.coll}
\alias{pc.cons.intern}
\alias{checkTriple}
\alias{faith.check}
\alias{triple2numb}
\alias{updateList}
\alias{minUncovPdPath}
\alias{minUncovCircPath}
\alias{find.min.discr.path}
\alias{qreach}
\alias{pdsep}
\alias{udag2pag}
\alias{find.unsh.triple}
\alias{rfci.vstructures}
\alias{dep.triple}
\alias{CheckEdges}
\alias{udag2apag}
\alias{ancTS}
\alias{skeleton.dag2pag}
\alias{new.ord}
\alias{VisibleEdge}
\alias{possibleDe}
\alias{dreach}
\alias{my.SpecialDag}
\alias{caus.inf}
%-> *do* document or confirm current status (FIXME):
\alias{gAlgo-class}
%
\title{Internal Pcalg Functions}
\description{
  Internal functions of package \pkg{pcalg}
}
\usage{
wgtMatrix(g, transpose=TRUE)
adj.check(gm,x)
amat2dag(amat)
find.sink(gm)
}
\details{
  These are not to be called by the user.
}
\keyword{ internal }
